requirements.txt
# Core deep learning
torch>=2.0
monai>=1.3
nnunetv2>=2.0

# Medical image I/O
SimpleITK>=2.2

# PI-CAI evaluation toolkit
picai_eval>=1.4
picai_baseline>=0.3
report_guided_annotation>=0.2

# Numerics / utilities
numpy>=1.23
einops>=0.6
pyyaml>=6.0
tqdm>=4.65
matplotlib>=3.6

# GPU monitoring (optional, used by src.utils.gpu_monitor)
pynvml>=11.5
